requirements.txt
streamlit
PyMuPDF
python-docx
pdfminer.six
